+++ example/systemjs.config.js
@@ -9,6 +9,7 @@
         "@angular/animations": "npm:@angular/animations/bundles/animations.umd.js",
         "@angular/animations/browser": "npm:@angular/animations/bundles/animations-browser.umd.js",
         "@angular/animations/testing": "npm:@angular/animations/bundles/animations-browser-testing.umd.js",
+        "@angular/cdk": "npm:@angular/cdk/bundles/cdk.umd.js",
         "@angular/common": "npm:@angular/common/bundles/common.umd.js",
         "@angular/common/testing": "npm:@angular/common/bundles/common-testing.umd.js",
         "@angular/compiler": "npm:@angular/compiler/bundles/compiler.umd.js",